requirements.txt
flask
gunicorn
python-docx
openai
PyMuPDF
Pillow
